+++ spec/features/staff/users_can_edit_an_organisation_spec.rb
@@ -21,8 +21,9 @@ RSpec.feature "Users can edit organisations" do
 
     visit organisation_path(beis_organisation)
     click_link I18n.t("page_title.organisation.index")
-    click_link another_organisation.name
-    click_link I18n.t("page_content.organisation.button.edit")
+    within("##{another_organisation.id}") do
+      click_link I18n.t("generic.link.edit")
+    end
 
     expect(page).to have_content(I18n.t("page_title.organisation.edit"))
     fill_in "organisation[name]", with: ""
@@ -37,8 +38,9 @@ RSpec.feature "Users can edit organisations" do
 
     click_link I18n.t("page_title.organisation.index")
 
-    click_link another_organisation.name
-    click_link I18n.t("page_content.organisation.button.edit")
+    within("##{another_organisation.id}") do
+      click_link I18n.t("generic.link.edit")
+    end
 
     expect(page).to have_content(I18n.t("page_title.organisation.edit"))
     fill_in "organisation[name]", with: "My New Organisation"